makde video detection component
